Complete rewrite
Closes #2
Uses PKHeX.Core dll for meat of logic, the UI is just to specify
destination type and configure conversion default values.

Bottlenecks are large counts of files as the read/write speed of the
file location.

For unsupported transfer methods, the user can opt to use a transfer
method driven by Reflection to transfer as many properties as possible
to the destination PKM, thus enabling backwards/incompatible transfer
routes.

UI locks up on large counts of files, potential for bettering that
handling.